Refactor Bricklayer: extract shared styles, Vec3Input, emitter presets
Extract duplicated panel styles into shared panelStyles (styles/panel.ts),
Vec3Input into a shared component (components/Vec3Input.tsx), and emitter
presets into data/emitterPresets.ts. Update all 11 panel files to import
from shared modules. Delete unused SceneTreePanel.tsx and Inspector.tsx.
