+++ tools/apps/bricklayer/src/panels/LightsTab.tsx
@@ -2,28 +2,9 @@ import React from 'react';
 import { NumberInput } from '../components/NumberInput.js';
 import { useSceneStore } from '../store/useSceneStore.js';
 import type { StaticLight } from '../store/types.js';
+import { panelStyles } from '../styles/panel.js';
 
-const styles: Record<string, React.CSSProperties> = {
-  section: { display: 'flex', flexDirection: 'column', gap: 8, marginBottom: 16 },
-  label: { fontSize: 11, color: '#888', textTransform: 'uppercase' as const, letterSpacing: 1 },
-  row: { display: 'flex', alignItems: 'center', gap: 8 },
-  input: { flex: 1, padding: '3px 5px', fontSize: 12 },
-  btn: {
-    padding: '4px 10px', border: '1px solid #555', borderRadius: 4,
-    background: '#3a3a6a', color: '#ddd', cursor: 'pointer', fontSize: 12,
-  },
-  btnDanger: {
-    padding: '4px 10px', border: '1px solid #c33', borderRadius: 4,
-    background: '#4a2020', color: '#faa', cursor: 'pointer', fontSize: 12,
-  },
-  item: {
-    padding: 8, border: '1px solid #444', borderRadius: 4, background: '#22223a',
-    display: 'flex', flexDirection: 'column', gap: 6,
-  },
-  itemSelected: {
-    borderColor: '#77f',
-  },
-};
+const styles = { ...panelStyles };
 
 function LightEditor({ light }: { light: StaticLight }) {
   const updateLight = useSceneStore((s) => s.updateLight);
